Create FilesWithDiff component
