+++ src/content/FilesWithDiff/FilesWithDiff.module.scss
@@ -1,0 +1,14 @@
+@import "../../shared-component-styles.scss";
+
+.metadata {
+  color: #6a737d;
+  font-weight: bold;
+}
+
+.added {
+  color: #22863a;
+}
+
+.removed {
+  color: #cb2431;
+}
