+++ src/content/FilesWithDiff/FilesWithDiff.tsx
@@ -1,0 +1,84 @@
+import { Text } from "@primer/react";
+import React, { useState } from "react";
+import { cns } from "ts-type-safe";
+import { Files } from "../types";
+import styles from "./FilesWithDiff.module.scss";
+
+type Props = {
+  prTitle: string;
+  files: Files;
+};
+
+export const FilesWithDiff: React.FC<Props> = ({ files, prTitle }) => {
+  const [open, openSet] = useState<string>();
+  return (
+    <>
+      {files.map((file, index) => (
+        <React.Fragment key={file.filename}>
+          <Text
+            as="li"
+            onMouseEnter={() => openSet(getKey(prTitle, index))}
+            onMouseLeave={() => openSet(undefined)}
+          >
+            {file.filename}
+          </Text>
+          {/* only text-based files have a patch */}
+          {file.patch && (
+            <div
+              className={cns(
+                styles.popup,
+                open === getKey(prTitle, index) && styles.popup__hovered
+              )}
+            >
+              <pre>
+                <code>{formatPatch(file.patch)}</code>
+              </pre>
+            </div>
+          )}
+        </React.Fragment>
+      ))}
+    </>
+  );
+};
+
+function getKey(prTitle: string, index: number) {
+  return `${prTitle}-${index}`;
+}
+
+// Function to escape HTML and highlight patch content
+const formatPatch = (patch: string | undefined) => {
+  if (!patch) return;
+
+  // Split lines and wrap each in appropriate JSX with highlights
+  return patch.split("\n").map((line, index) => {
+    if (line.startsWith("@@")) {
+      // Separate the metadata and code parts on the line
+      const [, metadata, code] = line.split("@@ ");
+      return (
+        <React.Fragment key={index}>
+          {/* Display metadata part */}
+          <div className={styles.metadata}>{metadata}</div>
+          {/* Display code part (if it exists) on a new line */}
+          {code ? <div>{code}</div> : null}
+        </React.Fragment>
+      );
+    } else if (line.startsWith("+")) {
+      // Added line
+      return (
+        <div key={index} className={styles.added}>
+          {line}
+        </div>
+      );
+    } else if (line.startsWith("-")) {
+      // Removed line
+      return (
+        <div key={index} className={styles.removed}>
+          {line}
+        </div>
+      );
+    } else {
+      // Regular code line
+      return <div key={index}>{line}</div>;
+    }
+  });
+};
