+++ src/content/FilesWithDiff/index.ts
@@ -1,0 +1,1 @@
+export { FilesWithDiff } from "./FilesWithDiff";
