+++ src/content/PrFilesSearch/PrFilesSearch.module.scss
@@ -12,16 +12,3 @@
 .prFiles:not(:first-child) {
   margin-top: 1rem;
 }
-
-.metadata {
-  color: #6a737d;
-  font-weight: bold;
-}
-
-.added {
-  color: #22863a;
-}
-
-.removed {
-  color: #cb2431;
-}
