+++ src/content/types.ts
@@ -1,0 +1,4 @@
+import { RestEndpointMethodTypes } from "@octokit/rest";
+
+export type Files =
+  RestEndpointMethodTypes["pulls"]["listFiles"]["response"]["data"];
